Create `JwtStrategy` to extract and decode token
