+++ src/auth/strategies/index.ts
@@ -1,0 +1,2 @@
+export { JwtStrategy } from './jwt.strategy';
+export { LocalStrategy } from './local.strategy';
